Rename PreGame and GameManager
